convert home.js to home.tsx
