+++ package.json
@@ -71,6 +71,7 @@
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.6",
+    "@types/recharts": "^1.8.16",
     "@types/styled-components": "^5.1.4",
     "axios": "^0.19.2",
     "babel-preset-es2015": "^6.24.1",
