+++ src/reducers/accountMap.ts
@@ -1,7 +1,9 @@
+import { AccountMapAction, AccountMap } from '../types'
+
 export const ACCOUNTMAP_UPDATE = 'ACCOUNTMAP_UPDATE';
 export const ACCOUNTMAP_SET = 'ACCOUNTMAP_SET';
 
-export const accountMapReducer = (state, action) => {
+export const accountMapReducer = (state: AccountMap, action: AccountMapAction) => {
   if (action.type === ACCOUNTMAP_UPDATE) {
     return {
       ...state,
