+++ src/types/declarations/unchained-bitcoin/index.d.ts
@@ -1,0 +1,3 @@
+declare module 'unchained-bitcoin' {
+  declare function satoshisToBitcoins(n: number): number
+}
